# api/openapi.yaml
openapi: 3.1.0
info:
  title: DokoSoko Control Plane API
  version: 4.0.0
  description: >-
    Control-plane contract for DokoSoko's core documentation, SDK references, credentials, recipes, reviewed tools, native tools, and MCP connector. Cookie-authenticated mutations
    also require CSRF validation. Public MCP is optional and read-only.
servers:
  - url: https://dokosoko.example.com
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /api/v1/identity-provider:
    get:
      operationId: getIdentityProvider
      responses:
        '200':
          description: OIDC customer identity configuration, or an explicit unconfigured setup shape containing the exact callback URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityProvider'
    put:
      operationId: replaceIdentityProvider
      description: Saves a disabled OIDC draft. State is server-controlled; every save creates a new revision and immediately invalidates OAuth artifacts bound to older revisions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceIdentityProviderRequest'
      responses:
        '200':
          description: Identity provider configuration saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityProvider'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
    delete:
      operationId: disconnectIdentityProvider
      description: >-
        Permanently disconnects a disabled OIDC configuration at one exact revision. Deletes the provider config, client credential, provider tests, and product OAuth artifacts
        with their delegated bearer secrets. Customer accounts and append-only audit history are intentionally preserved. Disable first so runtime access stops as a separate
        explicit action.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdentityProviderRevisionRequest'
      responses:
        '200':
          description: Clean explicit unconfigured setup shape.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityProvider'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
  /oauth/register:
    post:
      security: []
      operationId: registerOAuthClient
      description: >-
        Registers a public PKCE client under RFC 7591. Identical normalized metadata returns the same durable client identifier, making network retries safe. No client secret is
        issued.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthClientRegistrationRequest'
      responses:
        '201':
          description: Idempotent public-client registration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClientRegistration'
        '400':
          $ref: '#/components/responses/OAuthError'
        '404':
          $ref: '#/components/responses/OAuthError'
        '429':
          $ref: '#/components/responses/OAuthRateLimit'
        '503':
          $ref: '#/components/responses/OAuthError'
  /oauth/authorize:
    get:
      security: []
      operationId: oauthAuthorize
      description: >-
        Starts authorization-code OAuth using a registered public client or HTTPS client metadata document, mandatory RFC 8707 resource binding, PKCE S256, upstream vendor OIDC,
        and fail-closed access evaluation.
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            const: code
        - name: client_id
          in: query
          required: true
          schema:
            type: string
            description: Durable RFC 7591 client identifier or HTTPS URL of a client metadata document.
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: resource
          in: query
          required: true
          schema:
            type: string
            format: uri
            description: Exact Private MCP resource URL advertised by protected-resource metadata.
        - name: scope
          in: query
          required: false
          schema:
            type: string
            default: mcp:private
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            const: S256
      responses:
        '302':
          description: Redirect to the configured vendor IdP.
        '400':
          $ref: '#/components/responses/OAuthError'
  /oauth/callback:
    get:
      security: []
      operationId: oauthCallback
      description: >-
        For public OAuth states, verifies upstream state, nonce and ID token, resolves vendor access fail-closed, creates or resolves the durable customer account, then redirects
        with a one-time code. Prefixed administration-test states stop after OIDC, audience/client, nonce, issuer, and customer-claim verification and redirect to the fixed
        same-origin Identity page without retaining tokens or creating runtime identity data.
      responses:
        '302':
          description: Exact registered client redirect.
        '303':
          description: >-
            Fixed same-origin Identity-page redirect carrying identity_test_id for a stored result or
            identity_test_error=invalid_or_expired when the callback transaction cannot be claimed.
        '401':
          $ref: '#/components/responses/OAuthError'
  /oauth/token:
    post:
      security: []
      operationId: oauthToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: false
              properties:
                grant_type:
                  const: authorization_code
                code:
                  type: string
                code_verifier:
                  type: string
                client_id:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
                resource:
                  type: string
                  format: uri
              required:
                - grant_type
                - code
                - code_verifier
                - client_id
                - redirect_uri
                - resource
      responses:
        '200':
          description: Short-lived opaque DokoSoko access token bound to the exact Private MCP resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthTokenResponse'
        '400':
          $ref: '#/components/responses/OAuthError'
        '502':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /mcp:
    parameters:
      - $ref: '#/components/parameters/MCPProtocolVersion'
      - $ref: '#/components/parameters/MCPMethod'
      - $ref: '#/components/parameters/MCPName'
    post:
      security:
        - dokoBearer: []
      operationId: privateDeploymentMCP
      description: Canonical authenticated Stateless MCPv2 endpoint for the singleton Deployment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONRPCRequest'
      responses:
        '200':
          description: JSON-RPC result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSONRPCResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dokosoko_session
    bearerAuth:
      type: http
      scheme: bearer
      description: Administrator automation token in development only; production uses the root session cookie.
    setupToken:
      type: http
      scheme: bearer
      description: One-time deployment setup token.
    dokoBearer:
      type: http
      scheme: bearer
      description: Opaque, short-lived DokoSoko OAuth access token bound to one exact Private MCP resource.
  parameters:
...
    MCPProtocolVersion:
      name: MCP-Protocol-Version
      in: header
      required: true
      schema:
        type: string
        const: '2026-07-28'
    MCPMethod:
      name: Mcp-Method
      in: header
      required: true
      description: Must exactly match the JSON-RPC method.
      schema:
        type: string
        enum:
          - server/discover
          - tools/list
          - tools/call
    MCPName:
      name: Mcp-Name
      in: header
      required: false
      description: Required for tools/call and must exactly match params.name.
      schema:
        type: string
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    OAuthError:
      description: OAuth 2.0 error response.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              error:
                type: string
              error_description:
                type: string
            required:
              - error
              - error_description
    OAuthRateLimit:
      description: OAuth client registration rate limit exceeded. Retrying after the advertised delay is safe.
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              error:
                const: temporarily_unavailable
              error_description:
                type: string
            required:
              - error
              - error_description
...
  schemas:
...
    ReplaceIdentityProviderRequest:
      type: object
      additionalProperties: false
      properties:
        provider:
          type: string
          const: oidc
          description: Explicit fixed provider discriminator for standards-based OpenID Connect.
        issuer:
          type: string
          format: uri
          description: >-
            Exact HTTPS OIDC issuer, including any path, port, or trailing slash. HTTP is accepted only for
            localhost development.
        client_id:
          type: string
          minLength: 1
        client_secret:
          type: string
          format: password
          writeOnly: true
          description: Required for initial configuration; omit or send empty to retain the current encrypted secret.
        scopes:
          type: array
          items:
            type: string
          default:
            - openid
            - profile
            - email
        audience:
          type: string
          description: Optional provider-specific authorization-request audience. Omitted upstream when empty.
        oauth_resource:
          type: string
          format: uri
          description: Optional absolute RFC 8707 resource identifier sent to the upstream authorization server; independent of the authorization API origin and not dereferenced by DokoSoko.
        customer_account_claim:
          type: string
          minLength: 1
          description: Required string ID-token claim that identifies the customer account; no implicit claim is assumed.
        installation_claim:
          type: string
          description: Optional trusted OIDC claim matched to a registered product installation.
        authorization_api_origin:
          description: >-
            Credential-free HTTPS origin used only with the delegated user token. Loopback and *.localhost HTTP origins with explicit ports are accepted only for local development.
            DokoSoko appends the normative access-evaluation path and rejects customer tool endpoints on other origins.
          type: string
          format: uri
        revision:
          type: integer
          format: int64
          minimum: 0
      required:
        - provider
        - issuer
        - client_id
        - customer_account_claim
        - authorization_api_origin
        - revision
    IdentityProviderRevisionRequest:
      type: object
      additionalProperties: false
      properties:
        revision:
          type: integer
          format: int64
          minimum: 1
      required:
        - revision
...
    IdentityProviderTest:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - pending
            - passed
            - failed
            - expired
        configuration_revision:
          type: integer
          format: int64
          minimum: 1
        authorization_url:
          type: string
          format: uri
          description: Returned only by the begin operation and never persisted.
        failure_code:
          type: string
          description: Stable non-secret lifecycle failure code. client_authentication_unsupported means discovery advertised neither client_secret_basic nor client_secret_post.
        issuer:
          type: string
          format: uri
        customer_id:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
      required:
        - id
        - status
        - configuration_revision
        - created_at
        - expires_at
    IdentityProvider:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        organisation_id:
          type: string
        deployment_id:
          type: string
        provider:
          type: string
          const: oidc
        configured:
          type: boolean
        credential_present:
          type: boolean
        callback_url:
          type: string
          format: uri
        access_evaluation_url:
          type: string
          description: Exact delegated evaluation URL, or empty before an API origin is configured.
        issuer:
          type: string
          description: Empty only in the first-run unconfigured response.
        client_id:
          type: string
        scopes:
          type: array
          items:
            type: string
        audience:
          type: string
        oauth_resource:
          type: string
          format: uri
        customer_account_claim:
          type: string
        installation_claim:
          type: string
        authorization_api_origin:
          type: string
        state:
          type: string
          enum:
            - active
            - disabled
        revision:
          type: integer
          format: int64
          minimum: 0
        last_test:
          $ref: '#/components/schemas/IdentityProviderTest'
        created_at:
          type: string
          format: date-time
          description: Present only after a configuration has been saved.
        updated_at:
          type: string
          format: date-time
          description: Present only after a configuration has been saved.
      required:
        - organisation_id
        - deployment_id
        - provider
        - configured
        - credential_present
        - callback_url
        - access_evaluation_url
        - issuer
        - client_id
        - scopes
        - audience
        - oauth_resource
        - customer_account_claim
        - installation_claim
        - authorization_api_origin
        - state
        - revision
...
    OAuthClientRegistrationRequest:
      type: object
      additionalProperties: true
      properties:
        redirect_uris:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uri
        client_name:
          type: string
          maxLength: 200
        token_endpoint_auth_method:
          type: string
          const: none
          default: none
        grant_types:
          type: array
          maxItems: 1
          items:
            const: authorization_code
        response_types:
          type: array
          maxItems: 1
          items:
            const: code
        scope:
          type: string
          const: mcp:private
      required:
        - redirect_uris
    OAuthClientRegistration:
      type: object
      additionalProperties: false
      properties:
        client_id:
          type: string
        client_id_issued_at:
          type: integer
          format: int64
        client_name:
          type: string
        redirect_uris:
          type: array
          minItems: 1
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            format: uri
        token_endpoint_auth_method:
          type: string
          const: none
        grant_types:
          type: array
          items:
            const: authorization_code
        response_types:
          type: array
          items:
            const: code
        scope:
          type: string
          const: mcp:private
      required:
        - client_id
        - client_id_issued_at
        - client_name
        - redirect_uris
        - token_endpoint_auth_method
        - grant_types
        - response_types
        - scope
...
    OAuthTokenResponse:
      type: object
      additionalProperties: false
      properties:
        access_token:
          type: string
          description: Opaque resource-bound DokoSoko access token.
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          minimum: 1
        scope:
          type: string
      required:
        - access_token
        - token_type
        - expires_in
        - scope
    JSONRPCID:
      type:
        - string
        - integer
        - 'null'
...
    MCPRequestMeta:
      type: object
      additionalProperties: true
      properties:
        io.modelcontextprotocol/protocolVersion:
          type: string
          const: '2026-07-28'
      required:
        - io.modelcontextprotocol/protocolVersion
    JSONRPCError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: integer
        message:
          type: string
        data: {}
      required:
        - code
        - message
    JSONRPCResponse:
      type: object
      additionalProperties: false
      properties:
        jsonrpc:
          const: '2.0'
        id:
          $ref: '#/components/schemas/JSONRPCID'
        result: {}
        error:
          $ref: '#/components/schemas/JSONRPCError'
      required:
        - jsonrpc
        - id
      oneOf:
        - required:
            - result
          not:
            required:
              - error
        - required:
            - error
          not:
            required:
              - result
...
    JSONRPCRequest:
      type: object
      additionalProperties: false
      properties:
        jsonrpc:
          const: '2.0'
        id:
          $ref: '#/components/schemas/JSONRPCID'
        method:
          type: string
          enum:
            - server/discover
            - tools/list
            - tools/call
            - resources/list
            - resources/read
        params:
          type: object
          additionalProperties: true
          properties:
            _meta:
              $ref: '#/components/schemas/MCPRequestMeta'
          required:
            - _meta
      required:
        - jsonrpc
        - method
        - params
      allOf:
        - if:
            properties:
              method:
                const: tools/call
          then:
            properties:
              params:
                $ref: '#/components/schemas/JSONRPCToolCallParams'
            required:
              - params
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - 'null'
              additionalProperties: true
          required:
            - code
            - message
      required:
        - error
...